First stab at adding APIs introduced in Xcode 10.2

+++ pyobjc-framework-Cocoa/Modules/_Foundation_netservice.m
@@ -92,7 +92,7 @@ static PyObject* call_NSNetService_addresses(
     PyObject* result;
     struct objc_super super;
     NSArray* res;
-    int len, i;
+    NSInteger len, i;
     NSData* item;
 
     if (!PyArg_ParseTuple(arguments, "")) {
